media: radio: s610: fix build errors
kernel/samsung/exynos850/drivers/media/radio/s610/radio-s610.c:1663:5: error: cast to smaller integer type 'unsigned int' from 'struct device *' [-Werror,-Wpointer-to-int-cast]
                                (unsigned int)radio->dev,
                                ^~~~~~~~~~~~~~~~~~~~~~~~
kernel/samsung/exynos850/drivers/media/radio/s610/radio-s610.c:1791:5: error: cast to smaller integer type 'unsigned int' from 'struct device *' [-Werror,-Wpointer-to-int-cast]
                                (unsigned int)radio->dev,
                                ^~~~~~~~~~~~~~~~~~~~~~~~
2 errors generated.

+++ drivers/media/radio/s610/radio-s610.c
@@ -1607,7 +1607,7 @@ static int s610_radio_fops_open(struct file *file)
 			}
 
 			abox_request_cpu_gear_sync(radio->a_dev,
-				dev_get_drvdata(radio->a_dev), (unsigned int)radio->dev, ABOX_CPU_GEAR_MAX, "FM");
+				dev_get_drvdata(radio->a_dev), (unsigned int)(uintptr_t)radio->dev, ABOX_CPU_GEAR_MAX, "FM");
 		}
 #endif /* USE_AUDIO_PM */
 
@@ -1725,7 +1725,7 @@ static int s610_radio_fops_release(struct file *file)
 #ifdef USE_AUDIO_PM
 		if (radio->a_dev) {
 			abox_request_cpu_gear_sync(radio->a_dev,
-				dev_get_drvdata(radio->a_dev), (unsigned int)radio->dev, ABOX_CPU_GEAR_MIN, "FM");
+				dev_get_drvdata(radio->a_dev), (unsigned int)(uintptr_t)radio->dev, ABOX_CPU_GEAR_MIN, "FM");
 
 			pm_runtime_put_sync(radio->a_dev);
 		}
